Refactoring tests structure.
